+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix compilation on freeBSD
 - fix `:filesystems` view not listing disks whose mount point has a space character
 - fix panic on searching `cr/.*` if a file starts with an empty line - Fix #406
+- fix preview of linux pseudo-files
 
 <a name="v1.6.0"></a>
 ### v1.6.0 - 2021-06-16